pyproject.toml
[project]
name = "api-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic==1.10.9",
    "pyyaml==6.0",
    "requests==2.31.0",
    "ruff>=0.7.0",
]

[tool.uv]
dev-dependencies = [
    "black>=24.10.0",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # PyLint
    "PL",
    # flake8-use-pathlib 
    "PTH"
]
